move HeroContentContainer to components
