+++ src/components/Hero/Hero.module.css
@@ -4,97 +4,8 @@
   border-radius: var(--border-radius-small, 20px);
 }
 
-.heroContentContainer {
-  padding-top: 194px;
-  padding-bottom: 338px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  gap: 20px;
-}
-
-.heroTitle {
-  width: 326px;
-  display: inline-block;
-  margin: 0 auto;
-
-  font-size: 40px;
-  font-weight: 800;
-  line-height: 1;
-  text-align: center;
-  color: var(--color-white, #ffffff);
-  text-transform: uppercase;
-}
-
-.heroSubtitle {
-  width: 326px;
-  display: inline-block;
-  margin: 0 auto;
-
-  font-size: 14px;
-  line-height: 1.42;
-  text-align: center;
-
-  color: #ffffffcc;
-}
-
-.heroSectionButton {
-  display: inline-block;
-  padding: 12px 24px;
-  margin: 0 auto;
-  border-radius: var(--border-radius-regular, 30px);
-  background-color: transparent;
-
-  font-size: 14px;
-  font-weight: 700;
-  line-height: 1.42;
-  text-align: center;
-  color: var(--color-white, #ffffff);
-  text-transform: uppercase;
-
-  border: 1px solid #ffffffe5;
-}
-
 @media screen and (min-width: 768px) {
   .heroSection {
     border-radius: var(--border-radius-regular, 30px);
-  }
-  .heroContentContainer {
-    padding-top: 218px;
-    padding-bottom: 450px;
-    gap: 40px;
-  }
-
-  .heroTitle {
-    width: 736px;
-
-    font-size: 70px;
-    line-height: 1;
-  }
-
-  .heroSubtitle {
-    width: 576px;
-
-    font-size: var(--font-size-body, 16px);
-    line-height: var(--line-height-body, 24px);
-  }
-
-  .heroSectionButton {
-    font-size: var(--font-size-body, 16px);
-    line-height: var(--line-height-body, 24px);
-  }
-}
-
-@media screen and (min-width: 1440px) {
-  .heroContentContainer {
-    padding-top: 154px;
-    padding-bottom: 452px;
-  }
-
-  .heroTitle {
-    width: 890px;
-
-    font-size: 90px;
-    line-height: 1;
   }
 }
