+++ src/components/HeroContentContainer/index.js
@@ -1,0 +1,1 @@
+export * from "./HeroContentContainer";
